AI_Server/diseaseDetectionServer/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
google-generativeai
google-ai-generativelanguage
python-multipart
requests
pyrebase4
firebase_admin